+++ idea/testData/intentions/convertObjectLiteralToClass/objectLiteralNoCapture.kt
@@ -1,0 +1,7 @@
+open class K
+
+fun foo(n: Int) {
+    val x = <caret>object : K() {
+        fun bar() = 1
+    }
+}