+++ idea/testData/quickfix/modifiers/noAbstractForAnonymousObject.kt
@@ -1,6 +1,7 @@
 // "Make 'object : T {}' abstract" "false"
 // ACTION: Implement members
 // ACTION: Split property declaration
+// ACTION: Convert object literal to class
 // ERROR: Object must be declared abstract or implement abstract member public abstract fun foo(): Unit defined in T
 interface T {
     fun foo()